+++ lib/src/widgets/country_selector/country_selector_navigator.dart
@@ -13,6 +13,10 @@ abstract class CountrySelectorNavigator {
   final bool searchAutofocus;
   final TextStyle? subtitleStyle;
   final TextStyle? titleStyle;
+  final InputDecoration? searchBoxDecoration;
+  final TextStyle? searchBoxTextStyle;
+  final Color? searchBoxIconColor;
+  final ScrollPhysics? scrollPhysics;
 
   const CountrySelectorNavigator({
     this.countries,
@@ -24,6 +28,10 @@ abstract class CountrySelectorNavigator {
     required this.searchAutofocus,
     this.subtitleStyle,
     this.titleStyle,
+    this.searchBoxDecoration,
+    this.searchBoxTextStyle,
+    this.searchBoxIconColor,
+    this.scrollPhysics,
   });
 
   Future<Country?> navigate(BuildContext context);
@@ -43,6 +51,10 @@ abstract class CountrySelectorNavigator {
       searchAutofocus: searchAutofocus,
       subtitleStyle: subtitleStyle,
       titleStyle: titleStyle,
+      searchBoxDecoration: searchBoxDecoration,
+      searchBoxTextStyle: searchBoxTextStyle,
+      searchBoxIconColor: searchBoxIconColor,
+      scrollPhysics: scrollPhysics,
     );
   }
 
@@ -56,6 +68,12 @@ abstract class CountrySelectorNavigator {
     bool sortCountries,
     String? noResultMessage,
     bool searchAutofocus,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) = DialogNavigator._;
 
   const factory CountrySelectorNavigator.searchDelegate({
@@ -66,6 +84,12 @@ abstract class CountrySelectorNavigator {
     bool sortCountries,
     String? noResultMessage,
     bool searchAutofocus,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) = SearchDelegateNavigator._;
 
   const factory CountrySelectorNavigator.bottomSheet({
@@ -78,6 +102,10 @@ abstract class CountrySelectorNavigator {
     bool searchAutofocus,
     TextStyle? subtitleStyle,
     TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) = BottomSheetNavigator._;
 
   const factory CountrySelectorNavigator.modalBottomSheet({
@@ -89,6 +117,12 @@ abstract class CountrySelectorNavigator {
     bool sortCountries,
     String? noResultMessage,
     bool searchAutofocus,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) = ModalBottomSheetNavigator._;
 
   const factory CountrySelectorNavigator.draggableBottomSheet({
@@ -103,6 +137,12 @@ abstract class CountrySelectorNavigator {
     bool sortCountries,
     String? noResultMessage,
     bool searchAutofocus,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) = DraggableModalBottomSheetNavigator._;
 }
 
@@ -120,6 +160,12 @@ class DialogNavigator extends CountrySelectorNavigator {
     bool sortCountries = false,
     String? noResultMessage,
     bool searchAutofocus = kIsWeb,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) : super(
           countries: countries,
           favorites: favorites,
@@ -128,6 +174,12 @@ class DialogNavigator extends CountrySelectorNavigator {
           sortCountries: sortCountries,
           noResultMessage: noResultMessage,
           searchAutofocus: searchAutofocus,
+          subtitleStyle: subtitleStyle,
+          titleStyle: titleStyle,
+          searchBoxDecoration: searchBoxDecoration,
+          searchBoxTextStyle: searchBoxTextStyle,
+          searchBoxIconColor: searchBoxIconColor,
+          scrollPhysics: scrollPhysics,
         );
 
   @override
@@ -156,6 +208,12 @@ class SearchDelegateNavigator extends CountrySelectorNavigator {
     bool sortCountries = false,
     String? noResultMessage,
     bool searchAutofocus = kIsWeb,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) : super(
           countries: countries,
           favorites: favorites,
@@ -164,6 +222,12 @@ class SearchDelegateNavigator extends CountrySelectorNavigator {
           sortCountries: sortCountries,
           noResultMessage: noResultMessage,
           searchAutofocus: searchAutofocus,
+          subtitleStyle: subtitleStyle,
+          titleStyle: titleStyle,
+          searchBoxDecoration: searchBoxDecoration,
+          searchBoxTextStyle: searchBoxTextStyle,
+          searchBoxIconColor: searchBoxIconColor,
+          scrollPhysics: scrollPhysics,
         );
 
   @override
@@ -205,6 +269,10 @@ class BottomSheetNavigator extends CountrySelectorNavigator {
     bool searchAutofocus = kIsWeb,
     TextStyle? subtitleStyle,
     TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) : super(
           countries: countries,
           favorites: favorites,
@@ -215,6 +283,10 @@ class BottomSheetNavigator extends CountrySelectorNavigator {
           searchAutofocus: searchAutofocus,
           subtitleStyle: subtitleStyle,
           titleStyle: titleStyle,
+          searchBoxDecoration: searchBoxDecoration,
+          searchBoxTextStyle: searchBoxTextStyle,
+          searchBoxIconColor: searchBoxIconColor,
+          scrollPhysics: scrollPhysics,
         );
 
   @override
@@ -245,14 +317,27 @@ class ModalBottomSheetNavigator extends CountrySelectorNavigator {
     bool sortCountries = false,
     String? noResultMessage,
     bool searchAutofocus = kIsWeb,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) : super(
-            countries: countries,
-            favorites: favorites,
-            addSeparator: addSeparator,
-            showCountryCode: showCountryCode,
-            sortCountries: sortCountries,
-            noResultMessage: noResultMessage,
-            searchAutofocus: searchAutofocus);
+          countries: countries,
+          favorites: favorites,
+          addSeparator: addSeparator,
+          showCountryCode: showCountryCode,
+          sortCountries: sortCountries,
+          noResultMessage: noResultMessage,
+          searchAutofocus: searchAutofocus,
+          subtitleStyle: subtitleStyle,
+          titleStyle: titleStyle,
+          searchBoxDecoration: searchBoxDecoration,
+          searchBoxTextStyle: searchBoxTextStyle,
+          searchBoxIconColor: searchBoxIconColor,
+          scrollPhysics: scrollPhysics,
+        );
 
   @override
   Future<Country?> navigate(BuildContext context) {
@@ -287,14 +372,27 @@ class DraggableModalBottomSheetNavigator extends CountrySelectorNavigator {
     bool sortCountries = false,
     String? noResultMessage,
     bool searchAutofocus = kIsWeb,
+    TextStyle? subtitleStyle,
+    TextStyle? titleStyle,
+    InputDecoration? searchBoxDecoration,
+    TextStyle? searchBoxTextStyle,
+    Color? searchBoxIconColor,
+    ScrollPhysics? scrollPhysics,
   }) : super(
-            countries: countries,
-            favorites: favorites,
-            addSeparator: addSeparator,
-            showCountryCode: showCountryCode,
-            sortCountries: sortCountries,
-            noResultMessage: noResultMessage,
-            searchAutofocus: searchAutofocus);
+          countries: countries,
+          favorites: favorites,
+          addSeparator: addSeparator,
+          showCountryCode: showCountryCode,
+          sortCountries: sortCountries,
+          noResultMessage: noResultMessage,
+          searchAutofocus: searchAutofocus,
+          subtitleStyle: subtitleStyle,
+          titleStyle: titleStyle,
+          searchBoxDecoration: searchBoxDecoration,
+          searchBoxTextStyle: searchBoxTextStyle,
+          searchBoxIconColor: searchBoxIconColor,
+          scrollPhysics: scrollPhysics,
+        );
 
   @override
   Future<Country?> navigate(BuildContext context) {