+++ lib/src/widgets/country_selector/search_box.dart
@@ -3,11 +3,17 @@ import 'package:flutter/material.dart';
 class SearchBox extends StatelessWidget {
   final Function(String) onChanged;
   final bool autofocus;
+  final InputDecoration? decoration;
+  final TextStyle? style;
+  final Color? searchIconColor;
 
   const SearchBox({
     Key? key,
     required this.onChanged,
     required this.autofocus,
+    this.decoration,
+    this.style,
+    this.searchIconColor,
   }) : super(key: key);
 
   @override
@@ -17,20 +23,21 @@ class SearchBox extends StatelessWidget {
       child: TextField(
         autofocus: autofocus,
         onChanged: onChanged,
-        decoration: InputDecoration(
-          prefixIcon: Icon(
-            Icons.search,
-            color: Theme.of(context).brightness == Brightness.dark
-                ? Colors.white54
-                : Colors.black38,
-          ),
-          filled: true,
-          isDense: true,
-          border: OutlineInputBorder(
-            borderSide: BorderSide.none,
-            borderRadius: BorderRadius.circular(20),
-          ),
-        ),
+        style: style,
+        decoration: decoration ??
+            InputDecoration(
+              prefixIcon: Icon(
+                Icons.search,
+                color:
+                    searchIconColor ?? (Theme.of(context).brightness == Brightness.dark ? Colors.white54 : Colors.black38),
+              ),
+              filled: true,
+              isDense: true,
+              border: OutlineInputBorder(
+                borderSide: BorderSide.none,
+                borderRadius: BorderRadius.circular(20),
+              ),
+            ),
       ),
     );
   }